libpriv/util: Add date field in auto-versioning
This adds an optional date field to the prefix
passed by automatic_version_prefix. An example of specifying
the field is as follows:

10.<date: %Y%m%d>

And the fields progress like:

10.20181218.0
10.20181218.1
10.20181218.2
10.20181219.0

The date format is used to create a new "current date" string,
as long as it uses valid directives.

If there is a problem reading the given date format,
an error is given and the next version is returned as NULL.

If no <date: ...> tag is detected in the auto version prefix,
the same behavior as before (appending .1 and incrementing) occurs.

This may be helpful to avoid writing glue code to auto-update
the version if a date string in the commit version is desired.
Otherwise, --add-metadata-string=version= is an alternative for
complete customization.